Add optional arg, -P, to pass in prefix file
This commit updates UserCtrl to allow a new final, optional, argument -P. This
argument takes in a character, [file_prefix]. When passed to ModEM, ModEM will
set `ctrl % storeSolnsInFile` to `.true.` and will set `ctrl % prefix` to
[file_prefix].

This commit also updates the `search_args` do loop in `parseArgs` to allow it
to better handle additional final optional arguments. It should be easy now to
expand it to allow additional optional arguments. Just ensure you set
`optArgSize` correctly. For single, standalone options without an argument,
`optArgSize` should be 1, options with an argument should set `optArgSize` to
2.

This commit also updates the user guide with the details of this new optional
argument for inversion.