Create routines for reading and writing solnVectors_t
This commits creates three new routines in SolnSpace that allow the reading and
writing of solnVector_t types. The routines will read and write the active
transmitter and the active polarization of the passed in solnVector_t.

These routines will be used to save partial solutions, thus, perhaps they are a
bit limited in their interface. It may be appropriate to have a routine that
takes in desired iTx and polarization and allocates an solnVector_t, rather
than requiring it be already allocated and created. I'll think about that, but
for now this should be sufficient.