Update Worker_Job to take in user ctrl from main program
This commit makes it so that the subroutine `Worker_Job` has the initial
`userdef_control` that is specified at runtime. I'm surprised this wasn't
already done, but perhaps all important information is already present in the
worker jobs via initGlobalData.

Regardless, this change will make it so the Worker_Job has both the `prefix`
and `storeSolnsInFile` that can be specified by the -P optional argument.