Create an MPI Context Holder for keeping track of communicators
This commit creates an ModEM_mpi_context_t type which will be used to store and
pass around MPI communicators to the EsolnManager.

If it works well, perhaps in future merges it can be used to in other areas of
ModEM.

As of this commit, the ModEM_mpi_context_t is not used, but will be in a future
commit.