Add the ESolnManager module
This commits adds the EsolnManager to the code base. As of this commit, the
code does not make any calls to it, but will in subsequent commits.

It also adds in the ModEM_mpi_context_t, which will be used to store and pass
around MPI communicators to the EsolnManager.