Correct run-time error in write for GFortran-15
GFortran-15 apparent no longer approves of this write statement in UserCtrl and
throws a run-time error when printing out this error statement when writing:

`Fortran runtime error: Missing comma between descriptors`

I'm not sure if the previous format was standard complaint, but I'm inclined to
think it was not, but I am not sure why it worked for so long.

Regardless, the change in the format for this write statement matches the
expected output, i.e. the argument separated by a space.