Add in Liu et al 2024 and do various citation cleanup
This commit adds in Liu et al. 2024 to the citation, codemeta and README files.

As well, it adds a new paragraph with a warning about using the spherical
coordinates.

As well, it removes the Egbert & Kelbert 2012 PDF as it is already available
online. It does keep Kelbert et al 2014, as that PDF is not available. I added
a small link in the README's to link to this PDF. This was a recommendation
from folks at CIG.

Lastly, it adds a clickable DOI for Egbert & Kelbert for usability.